alarm tests: use ptr-sized types

+++ examples/tests/alarms/multi_alarm_dont_ignore_previous_overflow/main.c
@@ -5,8 +5,8 @@
 #include <libtock-sync/services/alarm.h>
 
 static void event_cb(uint32_t now, uint32_t expiration, void* ud) {
-  int i = (int)ud;
-  printf("%d %" PRIu32 " %" PRIu32 "\n", i, now, expiration);
+  uintptr_t i = (uintptr_t)ud;
+  printf("%zu %" PRIu32 " %" PRIu32 "\n", i, now, expiration);
 }
 
 int main(void) {
